Publish 'robotPose' in a format readable by AdvantageScope
To be readable by plotting and charting utilities, Pose2d information
needs to be published in a format compatible with Java Serialization.
This format is recognized by AdvantageScope and other FRC utilities.

The FRC-4121 robot's current pose estimate is now published as the
NetworkTable topic '/SmartDashboard/robotPose'.

Other poses of interest (such as destinations, location estimates of
field objects, etc.) can be published and plotted using 'robotPose' as
a design pattern.